add lock function and various other improvements

+++ scripts/create-gocryptfs.sh
@@ -5,6 +5,11 @@ if [[ $(</proc/sys/kernel/hostname) != "etrial" ]]; then
   exit
 fi
 
+if [[ $(whoami) != "etrial" ]]; then
+  echo "You should run this as the etrial user so that the resulting data is owned by the correct user."
+  exit
+fi
+
 if [[ "$1" == "purge" ]]; then
   fusermount -u /crypt
   rm -rf ~etrial/crypt
